Refactoring to adhere to eslint-config-athom, added optional parameter on capabilities option
The optional parameter in the capability specification in driver.js can
specify whether ZwaveDriver should accept a command class as being
variable for a specific capability. This prevents the driver to crash
when a device stops advertising a certain command class, e.g. command
class battery when a device is being powered by dc.