custom saveSettingMessage
There are certain devices that would really benefit of a custom save
Setting Message.
Like the Fibaro Keyfob,
who only wakes up enough in/for homey when using the NIF send sequence.
(which is different from what is in the manual)
And people... just don't read the readme file, this is a bit more in to
their face.

Since it's probably something only for z-wave,
I have choosen to put it in the zwave object,
and does work without having to re-pair the device.

I'll see if you agree on this idea, and/or if you would like to change
something.